Small Integer sin signo en profits de Playlist
A peticion del scrum manager

+++ database/migrations/2020_08_28_000625_create_playlists_table.php
@@ -17,7 +17,7 @@ class CreatePlaylistsTable extends Migration
             $table->id();
             $table->timestamps();
             $table->decimal('tier', 2, 1);    //Nivel del playlist
-            $table->unsignedTinyInteger('profits');  //Dinero que ha obtenido con la playlist
+            $table->unsignedSmallInteger('profits');  //Dinero que ha obtenido con la playlist
             $table->string('link_playlist');        //Link de spotify de la playlist
             $table->unsignedBigInteger('user_id');  //Musico que lanza la campaña
